block users from disabling the reboot and security preview notifications
This reverts commit 5fa7f8279c826c68f71c4bad356b681f4508e427. Conflicts are resolved, and the new
security preview notification is marked as non-blockable.

There is a case where I can reproduce the original issue of all the `NotificationChannel`s being
unblockable by doing the following:

- Revert the "refactor notification handling" commit, and
- Setting up conditions in Updater that would prevent a periodic update check after reboot

I reverted the refactor commit to try to reproduce the issue using the old behavior where the
`NotificationChannel` creation is done in the `onCreate` method of the `Service`. After the revert,
the new `NotificationChannel#setBlockable` calls wouldn't be active until a periodic or forced
update check happens (i.e. a `Service` is created)

This can happen quite easily, e.g.,

1. Someone can set it to require charging for periodic update checks
2. Then after update finishes, during the reboot, they unplug from charger (and maybe leave their
house)
3. With the new Updater version, the fixed notification permission will make all channels
non-blockable by default
4. A periodic check or a forced update check from the Updater settings is now required for the
`setBlockable(true)` calls to come into effect from the `Service`'s `onCreate` (one-time thing) to
opt in channels to being blockable.
5. Periodic check doesn't happen because phone is not charging, so notification channels remain
unblockable until they charge phone + have internet for the `Service` to run again
6. If they don't plug in their phone for a long time (not uncommon behavior), a user might've
noticed this issue of all channels being unblockable and reported it

Theoretically this bug, or at least this specific case of the bug, is fixed by the
"refactor notification handling" commit. `NotificationChannel`s would always be created (and thus
updated with the new `setBlockable` calls) on `Application `start instead of being limited to the
`Service#onCreate`.  The `Service` can be delayed for many reasons, so it's not an ideal place to
put `NotificationChannel` code.

It's unknown if there are other cases where all the channels become unblockable.

Co-authored-by: haok1402 <haok1402@gmail.com>
Test: Turning the reboot notification off will result in it enabled (and unblockable) after update
Test: Notification channels other than security preview and reboot preserve their user settings
after update, and they're blockable